eai_tutorial_code/requirements.txt
matplotlib~=3.4.2
IPython~=7.26.0
tensorboard
transformers
h5py
torch~=1.9.0
torchvision~=0.10.0
numpy~=1.20.3
Pillow~=8.3.1